voice_model_service/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
bibtexparser==2.0.0b8
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
clldutils==3.24.0
colorama==0.4.6
colorlog==6.9.0
constantly==23.10.4
cryptography==44.0.0
csvw==3.5.1
daphne==4.1.2
Django==5.1.4
djangorestframework==3.15.2
dlinfo==1.2.1
filelock==3.16.1
fsspec==2024.12.0
huggingface-hub==0.27.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
language-tags==1.2.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
phonemizer==3.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pylatexenc==2.10
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rdflib==7.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.22.3
safetensors==0.4.5
segments==2.2.1
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
soundfile==0.12.1
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
zope.interface==7.2